changed to tsx file && moved helper functions to util file
