--- src/views/Notifications/Home/_components/_utils.ts
@@ -1,36 +1,0 @@
-import {
-  GetNotifications_currentUser_notifications_notifications_activity_metaData_DailyDigestCompleteActivityMeta as DailyDigestCompleteActivityType,
-  GetNotifications_currentUser_notifications_notifications_activity_metaData_TaggedInDiscussionReplyActivityMeta as TaggedInDiscussionReplyActivityType,
-  GetNotifications_currentUser_notifications_notifications_activity_metaData_TaggedInPlanDiscussionActivityMeta as TaggedInDiscussionActivityType
-} from 'gql/gen/types/GetNotifications';
-
-// Type guard to check union type
-export const isTaggedInDiscussion = (
-  data:
-    | TaggedInDiscussionReplyActivityType
-    | TaggedInDiscussionActivityType
-    | DailyDigestCompleteActivityType
-): data is TaggedInDiscussionActivityType => {
-  /* eslint no-underscore-dangle: 0 */
-  return data.__typename === 'TaggedInPlanDiscussionActivityMeta';
-};
-
-export const isTaggedInDiscussionReply = (
-  data:
-    | TaggedInDiscussionReplyActivityType
-    | TaggedInDiscussionActivityType
-    | DailyDigestCompleteActivityType
-): data is TaggedInDiscussionReplyActivityType => {
-  /* eslint no-underscore-dangle: 0 */
-  return data.__typename === 'TaggedInDiscussionReplyActivityMeta';
-};
-
-export const isDailyDigest = (
-  data:
-    | TaggedInDiscussionReplyActivityType
-    | TaggedInDiscussionActivityType
-    | DailyDigestCompleteActivityType
-): data is DailyDigestCompleteActivityType => {
-  /* eslint no-underscore-dangle: 0 */
-  return data.__typename === 'DailyDigestCompleteActivityMeta';
-};
